% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_text_miscue.R
\name{chunk_text_miscue}
\alias{chunk_text_miscue}
\title{Chunk strings into begining middle end}
\usage{
chunk_text_miscue(text, miscue, ...)
}
\arguments{
\item{text}{The text field}

\item{miscue}{The miscue field}

\item{\ldots}{ignored}
}
\description{
Chunk strings into begining middle end
}
\examples{
chunk_text_miscue(miscue$text, miscue$miscue)
}
