% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_first_letter.R
\name{compare_first_letter}
\alias{compare_first_letter}
\title{Compare the first letter of chunked strings}
\usage{
compare_first_letter(text, miscue, chunk, ...)
}
\arguments{
\item{text}{The chunked text}

\item{miscue}{The chunked miscue}

\item{chunk}{A letter indicating if it's the beginning (b), middle (m), or end (e)
chunk.}

\item{\ldots}{ignored}
}
\description{
Compare the first letter of chunked strings
}
