% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_shape.R
\name{compare_shape}
\alias{compare_shape}
\title{Compare the shape of chunked strings}
\usage{
compare_shape(text, miscue, ...)
}
\arguments{
\item{text}{The chunked text}

\item{miscue}{The chunked miscue}

\item{\ldots}{ignored}
}
\description{
Compare the shape of chunked strings
}
