% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_size.R
\name{compare_size}
\alias{compare_size}
\title{Compare the size of chunked strings}
\usage{
compare_size(text, miscue, ...)
}
\arguments{
\item{text}{The chunked text}

\item{miscue}{The chunked miscue}

\item{\ldots}{ignored}
}
\description{
Compare the size of chunked strings
}
