% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicMiscue-package.R
\docType{data}
\name{miscue}
\alias{miscue}
\title{Miscues}
\format{A data frame with 21 rows and 2 variables}
\usage{
data(miscue)
}
\description{
A dataset containing miscues and text
}
\details{
\itemize{
  \item miscue. A Miscue
  \item text. A Text
}
}
\keyword{datasets}
